Refactor middleware interface
- Change the file name
- Change the way function is specified
