+++ app.ts
@@ -4,7 +4,7 @@ import path = require("path");
 import cookieParser = require("cookie-parser");
 import logger = require("morgan");
 import { fairbnRouter } from "./routes/fairbinden";
-import { HttpException } from "./interfaces/middlewareInterface";
+import { HttpException } from "./interfaces/middleware";
 
 export const app = express();
