+++ interfaces/middleware.ts
@@ -1,10 +1,8 @@
 import { Request, Response, NextFunction } from "express";
 
-export type MiddlewareFn = (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => void;
+export interface MiddlewareFn {
+  (req: Request, res: Response, next: NextFunction): void;
+}
 
 export interface HttpException extends Error {
   status: number;
